requirements.txt
google-cloud-bigquery==1.24.0
tensorflow==2.2.0
